Refactor: 이미지 import 위치 변경
